+++ src/main/java/com/simplilearn/fitnessclubautomation/controller/TrainerController.java
@@ -8,6 +8,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -22,18 +24,28 @@ public class TrainerController {
     }
 
     @RequestMapping(value = "/trainer/add", method = RequestMethod.POST)
-    private String addTrainer(@RequestParam(value = "trainer_name", required = true) String trainer_name,
+    private String addTrainer(ModelMap modelMap, HttpServletRequest request,
+                              @RequestParam(value = "trainer_name", required = true) String trainer_name,
                               @RequestParam(value = "trainer_age", required = true) int trainer_age,
                               @RequestParam(value = "trainer_gender", required = true) String trainer_gender,
                               @RequestParam(value = "trainer_experience", required = true) int trainer_experience,
                               @RequestParam(value = "trainer_address", required = true) String trainer_address) {
+        HttpSession session = request.getSession();
+        if (session.getAttribute("adminId") == null) {
+            return "redirect:/";
+        }
         Trainer tr = new Trainer(trainer_name, trainer_age, trainer_gender, trainer_experience, trainer_address);
         tr = trainerService.addTrainer(tr);
         return "redirect:/trainer";
     }
 
     @RequestMapping(value = "/trainer/edit-trainer/{trainer_id}")
-    private String editTrainer(ModelMap modelMap, @PathVariable Long trainer_id) {
+    private String editTrainer(ModelMap modelMap, HttpServletRequest request,
+                               @PathVariable Long trainer_id) {
+        HttpSession session = request.getSession();
+        if (session.getAttribute("adminId") == null) {
+            return "redirect:/";
+        }
         int[] experience = {1, 2, 3, 4, 5, 6, 7, 8, 9};
         modelMap.addAttribute("experience", experience);
         Trainer trainer = trainerService.getTrainer(trainer_id);
@@ -42,19 +54,29 @@ public class TrainerController {
     }
 
     @RequestMapping(value = "/trainer/edit", method = RequestMethod.POST)
-    private String editTrainer(@RequestParam(value = "trainer_id", required = true) Long trainer_id,
+    private String editTrainer(ModelMap modelMap, HttpServletRequest request,
+                               @RequestParam(value = "trainer_id", required = true) Long trainer_id,
                                @RequestParam(value = "trainer_name", required = true) String trainer_name,
                                @RequestParam(value = "trainer_age", required = true) int trainer_age,
                                @RequestParam(value = "trainer_gender", required = true) String trainer_gender,
                                @RequestParam(value = "trainer_experience", required = true) int trainer_experience,
                                @RequestParam(value = "trainer_address", required = true) String trainer_address) {
+        HttpSession session = request.getSession();
+        if (session.getAttribute("adminId") == null) {
+            return "redirect:/";
+        }
         Trainer tr = new Trainer(trainer_id, trainer_name, trainer_age, trainer_gender, trainer_experience, trainer_address);
         trainerService.addTrainer(tr);
         return "redirect:/trainer";
     }
 
     @RequestMapping(value = "/trainer/delete/{trainer_id}", method = RequestMethod.GET)
-    private String deleteTrainer(ModelMap modelMap, @PathVariable Long trainer_id) {
+    private String deleteTrainer(ModelMap modelMap, HttpServletRequest request,
+                                 @PathVariable Long trainer_id) {
+        HttpSession session = request.getSession();
+        if (session.getAttribute("adminId") == null) {
+            return "redirect:/";
+        }
         try {
             trainerService.deleteTrainer(trainer_id);
             modelMap.addAttribute("success", true);
@@ -68,7 +90,11 @@ public class TrainerController {
     }
 
     @RequestMapping(value = "/trainer")
-    private String getAllTrainers(ModelMap modelMap) {
+    private String getAllTrainers(ModelMap modelMap, HttpServletRequest request) {
+        HttpSession session = request.getSession();
+        if (session.getAttribute("adminId") == null) {
+            return "redirect:/";
+        }
         List<Trainer> trainers = new ArrayList<>();
         try {
             trainers = trainerService.getAllTrainers();
@@ -87,7 +113,11 @@ public class TrainerController {
     }
 
     @RequestMapping(value = "/trainer/{trainer_id}")
-    private String getTrainer(ModelMap modelMap, @PathVariable Long trainer_id) {
+    private String getTrainer(ModelMap modelMap, HttpServletRequest request, @PathVariable Long trainer_id) {
+        HttpSession session = request.getSession();
+        if (session.getAttribute("adminId") == null) {
+            return "redirect:/";
+        }
         Trainer trainer = new Trainer();
         try {
             trainer = trainerService.getTrainer(trainer_id);
